requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
langchain-community
beautifulsoup4
requests